Adapted tool/* to newly added tlspool_configvar() call; testdata generates properly
