+++ lib/Makefile
@@ -27,7 +27,7 @@ ifdef WINVER
 # TODO libtlspool_lidentry and libtlspool_pinentry
 libtlspool_OBJS = libtlspool.o
 else
-libtlspool_OBJS = libtlspool.o libtlspool_lidentry.o libtlspool_pinentry.o
+libtlspool_OBJS = libtlspool.o libtlspool_lidentry.o libtlspool_pinentry.o libtlspool_configvar.o
 endif
 
 OBJS = $(libtlspool_OBJS)
