+++ lib/libtlspool_pinentry.c
@@ -60,6 +60,9 @@ int tlspool_pin_service (char *path, uint32_t regflags, int responsetimeout_usec
 
 	/* Access the TLS Pool socket */
 	if (path == NULL) {
+		path = tlspool_configvar (NULL, "daemon_pidfile");
+	}
+	if (path == NULL) {
 		path = TLSPOOL_DEFAULT_SOCKET_PATH;
 	}
 	if (strlen (path) + 1 > sizeof (sun.sun_path)) {
