+++ tool/Makefile
@@ -19,9 +19,6 @@ PREFIX = /usr/local
 tlspool_LIB = ../lib/libtlspool.so
 tlspool_LDFLAGS = -L ../lib -ltlspool
 
-config_CFLAGS = -DCONFIG_PARSE_ONLY
-config_SRC = ../src/config.c
-
 donai_CFLAGS = -DDONAI_ONLY_TEXTPROC
 donai_SRC = ../src/donai.c
 
@@ -60,23 +57,23 @@ lidsel: lidsel.c $(tlspool_LIB)
 pinentry: pinentry.c $(tlspool_LIB)
 	gcc -std=gnu11 $(CFLAGS) -o "$@" "$<" $(tlspool_LDFLAGS)
 
-get_localid: get_localid.c $(config_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+get_localid: get_localid.c
+	gcc -std=gnu11 $(CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
-set_localid: set_localid.c $(config_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+set_localid: set_localid.c
+	gcc -std=gnu11 $(CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
-get_disclose: get_disclose.c $(config_SRC) $(donai_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(donai_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(donai_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+get_disclose: get_disclose.c $(donai_SRC)
+	gcc -std=gnu11 $(CFLAGS) $(donai_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(donai_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
-set_disclose: set_disclose.c $(config_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+set_disclose: set_disclose.c
+	gcc -std=gnu11 $(CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
-get_trust: get_trust.c $(config_SRC) $(donai_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(donai_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(donai_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+get_trust: get_trust.c $(donai_SRC)
+	gcc -std=gnu11 $(CFLAGS) $(donai_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(donai_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
-set_trust: set_trust.c $(config_SRC)
-	gcc -std=gnu11 $(CFLAGS) $(config_CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(config_SRC) $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
+set_trust: set_trust.c
+	gcc -std=gnu11 $(CFLAGS) $(BDB_CFLAGS) -o "$@" "$<" $(tlspool_LDFLAGS) $(BDB_LDFLAGS)
 
 pgp11_genkey: pgp11_genkey.c
 	gcc -std=gnu11 $(CFLAGS) $(pgp11genkey_CFLAGS) -o "$@" "$<" $(pgp11genkey_LIBS)
