+++ tool/set_trust.c
@@ -54,9 +54,9 @@ struct typemap_t typemap [] = {
 
 
 /* Setup and tear down management */
-int setup_management (DB_ENV **dbenv, DB_TXN **txn, DB **dbh) {
-	char *dbenv_dir = cfg_dbenv_dir ();
-	char *dbtad_fnm = cfg_db_trust ();
+int setup_management (char *cfgfile, DB_ENV **dbenv, DB_TXN **txn, DB **dbh) {
+	char *dbenv_dir = tlspool_configvar (cfgfile, "dbenv_dir");
+	char *dbtad_fnm = tlspool_configvar (cfgfile, "db_trust");
 	if (dbenv_dir == NULL) {
 		fprintf (stderr, "Please configure database environment directory\n");
 		return 0;
@@ -117,6 +117,7 @@ int main (int argc, char *argv []) {
 	DBT e_value;
 	int nomore;
 	int fd;
+	char *cfgfile;
 	//
 	// Sanity check
 	if ((argc < 5) || (argc > 6)) {
@@ -125,9 +126,10 @@ int main (int argc, char *argv []) {
 	}
 	//
 	// Initialise the modules taken from the src directory
-	parse_cfgfile (argv [1], 0);
+	;
 	//
 	// Prepare variables from arguments
+	cfgfile = argv [1];
 	if (argc >= 6) {
 		binfile = argv [5];
 	}
@@ -175,7 +177,7 @@ int main (int argc, char *argv []) {
 	}
 	//
 	// Now modify the matching entries
-	if (!setup_management (&dbenv, &txn, &dbh)) {
+	if (!setup_management (cfgfile, &dbenv, &txn, &dbh)) {
 		exit (1);
 	}
 	if (dbh->cursor (dbh, txn, &crs, 0) != 0) {
